requirements.txt
requests
python-telegram-bot
python-dotenv
starlette
uvicorn
redis